feat: session을 사용하여 유저 확인

+++ src/main/java/com/sparta/schedule_develop/service/CommentService.java
@@ -23,9 +23,9 @@ public class CommentService {
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
 
-    public CommentResponseDto save(CommentCreateRequestDto dto) {
+    public CommentResponseDto save(CommentCreateRequestDto dto, User loginUser) {
 
-        User user = userRepository.findById(dto.getUserId()).orElseThrow(() -> new IllegalArgumentException("해당 유저가 존재하지 않습니다."));
+        User user = userRepository.findById(loginUser.getId()).orElseThrow(() -> new IllegalArgumentException("해당 유저가 존재하지 않습니다."));
 
         Schedule schedule = scheduleRepository.findById(dto.getSchedulId()).orElseThrow(() -> new IllegalArgumentException("해당 할일이 존재하지 않습니다."));
 
@@ -47,8 +47,12 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentResponseDto updateById(Long id, CommentUpdateRequestDto dto) {
+    public CommentResponseDto updateById(Long id, CommentUpdateRequestDto dto, User loginUser) {
         Comment comment = commentRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 댓글이 존재하지 않습니다." + id));
+
+        if (!comment.getUser().getId().equals(loginUser.getId())) {
+            throw new IllegalStateException("댓글을 수정할 권한이 없습니다.");
+        }
 
         comment.update(dto.getContent());
 
@@ -56,8 +60,13 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteById(Long id) {
+    public void deleteById(Long id, User loginUser) {
         Comment comment = commentRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 댓글이 존재하지 않습니다." + id));
+
+        if (!comment.getUser().getId().equals(loginUser.getId())) {
+            throw new IllegalStateException("댓글을 삭제할 권한이 없습니다.");
+        }
+        
         commentRepository.delete(comment);
     }
